blogilistan laajennus, step10

+++ osa4/blogilista/controllers/blogs.js
@@ -101,8 +101,28 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const body = request.body
+  logger.info(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  //logger.info("Debug... Program gets to this line of code...")
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'error: token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  logger.info(user._id)
+  
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === decodedToken.id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'error: only user who added blog can remove it' })
+  }
+
+
+
+  
 })
